requirements.txt
flask
flask-login
werkzeug
mysql-connector-python